requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
cached-property==1.5.2
click==7.1.2
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
djangorestframework==3.12.4
gast==0.4.0
google-pasta==0.2.0
grpcio==1.37.0
gunicorn==20.1.0
h5py==3.1.0
importlib-metadata==3.10.0
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
nltk==3.6.1
numpy==1.19.5
Pillow==8.2.0
protobuf==3.15.8
psycopg2==2.8.6
pytz==2021.1
regex==2021.4.4
six==1.15.0
sqlparse==0.4.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tflearn==0.5.0
tqdm==4.60.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.1
